feat: Add itemNames, itemDescriptions, itemTradeFor, and itemImages to Mock.ts
