+++ sightingsMobileUI/src/screens/Mock.ts
@@ -24,3 +24,55 @@ export const userItems = [
       "https://images.pexels.com/photos/1506991/pexels-photo-1506991.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
   },
 ];
+
+export const itemNames = [
+  "Laptop",
+  "Smartphone",
+  "Headphones",
+  "Backpack",
+  "Water Bottle",
+  "Notebook",
+  "Pen",
+  "Sunglasses",
+  "Camera",
+  "Charger",
+];
+
+export const itemDescriptions = [
+  "A powerful device.",
+  "A handy gadget.",
+  "High-quality sound.",
+  "A sturdy backpack.",
+  "Stay hydrated.",
+  "Keep your notes.",
+  "For smooth writing.",
+  "Protect your eyes.",
+  "Capture moments.",
+  "Stay charged.",
+];
+
+export const itemTradeFor = [
+  "Tablet",
+  "Smartwatch",
+  "Speakers",
+  "Travel bag",
+  "Fitness tracker",
+  "E-book reader",
+  "Stylus pen",
+  "Hat",
+  "Lens",
+  "Power bank",
+];
+
+export const itemImages = [
+  "https://images.pexels.com/photos/18105/pexels-photo.jpg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/607812/pexels-photo-607812.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/3394651/pexels-photo-3394651.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/374898/pexels-photo-374898.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/416528/pexels-photo-416528.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/1187344/pexels-photo-1187344.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/158780/pexels-photo-158780.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/46710/pexels-photo-46710.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/167832/pexels-photo-167832.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+  "https://images.pexels.com/photos/399321/pexels-photo-399321.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2",
+];
