+++ sightingsMobileUI/src/screens/useProductList.ts
@@ -1,0 +1,37 @@
+import { useState } from "react";
+import { useDispatch } from "react-redux";
+import { v4 as uuidv4 } from "uuid";
+import { addItem, removeItem, UserItem } from "./Data/UserItems";
+
+const useProductList = (
+  itemNames: string[],
+  itemDescriptions: string[],
+  itemImages: string[],
+  itemTradeFor: string[]
+) => {
+  const [currentIndex, setCurrentIndex] = useState(0);
+  const dispatch = useDispatch();
+
+  const handleAddItem = () => {
+    const newItem: UserItem = {
+      id: uuidv4(),
+      name: itemNames[currentIndex],
+      description: itemDescriptions[currentIndex],
+      image: itemImages[currentIndex],
+      tradeFor: itemTradeFor[currentIndex],
+    };
+    dispatch(addItem(newItem));
+    setCurrentIndex((currentIndex + 1) % itemNames.length);
+  };
+
+  const handleRemoveItem = (id: string) => {
+    dispatch(removeItem(id));
+  };
+
+  return {
+    handleAddItem,
+    handleRemoveItem,
+  };
+};
+
+export default useProductList;
